Refactor

+++ BlueSheep/Common/Protocol/types/NetworkType.cs
@@ -4,7 +4,7 @@ namespace BlueSheep.Common.Protocol.Types
 {
     public abstract class NetworkType : MarshalByRefObject
     {
-        public abstract short TypeID { get; }
+        public abstract int TypeID { get; }
 
         public abstract void Serialize(IDataWriter writer);
         public abstract void Deserialize(IDataReader reader);
